refactor: extract large SwiftUI screens into sections and components
