+++ solidarity/Views/CardViews/OCRScanner/OCRScannerComponents.swift
@@ -1,0 +1,101 @@
+//
+//  OCRScannerComponents.swift
+//  solidarity
+//
+
+import AVFoundation
+import SwiftUI
+
+// MARK: - Extracted Field View
+
+struct ExtractedFieldView: View {
+    let label: String
+    let value: String
+    let confidence: Float?
+
+    var body: some View {
+        HStack {
+            VStack(alignment: .leading, spacing: 2) {
+                Text(label)
+                    .font(.caption)
+                    .foregroundColor(.secondary)
+
+                Text(value)
+                    .font(.body)
+            }
+
+            Spacer()
+
+            if let confidence = confidence {
+                ConfidenceBadge(confidence: confidence)
+            }
+        }
+    }
+}
+
+// MARK: - Confidence Badge
+
+struct ConfidenceBadge: View {
+    let confidence: Float
+
+    var body: some View {
+        let percentage = Int(confidence * 100)
+        let color: Color = {
+            if confidence >= 0.8 { return .green } else if confidence >= 0.6 { return .orange } else { return .red }
+        }()
+
+        Text("\(percentage)%")
+            .font(.caption)
+            .fontWeight(.semibold)
+            .padding(.horizontal, 8)
+            .padding(.vertical, 4)
+            .background(color.opacity(0.2))
+            .foregroundColor(color)
+            .cornerRadius(6)
+    }
+}
+
+// MARK: - Camera View
+
+struct CameraView: UIViewControllerRepresentable {
+    let onImageCaptured: (UIImage) -> Void
+
+    func makeUIViewController(context: Context) -> UIImagePickerController {
+        let picker = UIImagePickerController()
+        picker.sourceType = .camera
+        picker.delegate = context.coordinator
+        picker.allowsEditing = false
+        picker.cameraCaptureMode = .photo
+        picker.cameraDevice = .rear
+        picker.cameraFlashMode = .auto
+        return picker
+    }
+
+    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {}
+
+    func makeCoordinator() -> Coordinator {
+        Coordinator(onImageCaptured: onImageCaptured)
+    }
+
+    class Coordinator: NSObject, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
+        let onImageCaptured: (UIImage) -> Void
+
+        init(onImageCaptured: @escaping (UIImage) -> Void) {
+            self.onImageCaptured = onImageCaptured
+        }
+
+        func imagePickerController(
+            _ picker: UIImagePickerController,
+            didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]
+        ) {
+            if let image = info[.originalImage] as? UIImage {
+                onImageCaptured(image)
+            }
+            picker.dismiss(animated: true)
+        }
+
+        func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+            picker.dismiss(animated: true)
+        }
+    }
+}
